Combine redundant messageLinks tests
